requirements.txt
typer==0.21.0
psycopg2-binary==2.9.11
dbt-core==1.11.2
dbt-postgres==1.10.0
cryptography==46.0.3